==> 2.1- Data Preparation - Train.csv.ipynb <==
import pandas as pd
from matplotlib import pyplot as plt
import seaborn as sns
import numpy as np

import warnings
warnings.filterwarnings('ignore')

df = pd.read_csv("Data/TrainDataAnalysis.csv")

df.head()

df = df.drop(['Ticket'], axis = 1)

df.head()

df['Embarked'].unique()

df['Cabin_New'] = df['Cabin'].apply(lambda x: 
                     list(x) if type(x) != float else 
                     np.nan
                )

df.head()

df['Cabin_New_2'] = df['Cabin_New'].apply(lambda x: 
                         x[0] if type(x) == list else 
                         np.nan
                    )

df = df.drop(['Cabin', 'Cabin_New'], axis=1)

df.head()

df['Embarked'].unique()

df['Embarked'].value_counts()

df.isnull().sum()

df['Age_group'].fillna('[20, 40)', inplace=True)

df['Embarked'].fillna('S', inplace=True)

df['Cabin_New_2'].fillna('C', inplace=True)

df = df.drop(['Age'], axis=1)

df = df.drop(['Fare'], axis=1)

df['Sex_New'] = df['Sex'].apply(lambda x: 1 if x == 'male' else 0)


df.head()

df = df.drop(['Sex'], axis=1)

class OneStepFeatEng():
    def __init__(self, df):
        self.df = df
        
    # Drop columns
    def drop_column(self, column_names):
        self.df.drop(column_names, axis=1, inplace=True)
        return self.df.head()
    
    ## Separated cols
    def separated_cols(self, target_var):
        target_var = target_var
        features = [x for x in list(self.df.columns) if x != target_var]
        
        cat_cols = [column_name for column_name in self.df.columns if self.df[column_name].dtypes=='O']
        num_cols = [column_name for column_name in self.df.columns if self.df[column_name].dtypes!='O']
        
        return features, cat_cols, num_cols
    
    # Boolean encoding
    def boolean_encoding(self, column_name, true_val):
        self.df[column_name] = self.df[column_name].apply(lambda x: 1 if x == true_val else 0)
        return self.df.head()
    
    # Final columns
    def final_columns(self):
        return self.df.columns
    
    # Final dataframe
    def final_dataframe(self):
        return self.df

X = OneStepFeatEng(df)

features, cat_cols, num_cols = X.separated_cols('Survived')

X.final_dataframe().head()

features

cat_cols

num_cols

X.final_dataframe().to_csv("Data/TrainFeatureEngined.csv", index=False)
