==> 2.3- Data Preparation - All.ipynb <==
import pandas as pd
import numpy as np

Train =  pd.read_csv("Data/TrainFeatureEngined.csv")
Test =  pd.read_csv("Data/TestFeatureEngined.csv")

Train.head()

Train.shape

Test.shape

Train['SibSp'].unique()

Test.shape

class GetDummies():
    
    def __init__(self, train, test):
        self.train = train
        self.test = test
    
    # Dummy encoding    
    def dummy_encoding(self, column_name):
        self.train['train'] = 1
        self.test['train'] = 0
        
        combined = pd.concat([self.train, self.test])
        column_encoded = pd.get_dummies(combined[column_name])
        column_encoded.columns = ['{}_{}'.format(column_name, x) for x in column_encoded.columns]
        
        combined = pd.concat([combined, column_encoded], axis=1)
        
        train = combined[combined['train'] == 1]
        test = combined[combined['train'] == 0]
        
        train.drop(['train'], axis=1, inplace=True)
        test.drop(['train'], axis=1, inplace=True)
        
        return train, test

Train['train'] = 1
Test['train'] = 0

combined = pd.concat([Train, Test])

column_encoded = pd.get_dummies(combined['Pclass'])

combined = pd.concat([combined, column_encoded], axis=1)

Train = combined[combined['train'] == 1]
Test = combined[combined['train'] == 0]

Train.drop(['train'], axis=1, inplace=True)
Test.drop(['train'], axis=1, inplace=True)

Train.head()

Test.head()

Train['train'] = 1
Test['train'] = 0

combined = pd.concat([Train, Test])

column_encoded = pd.get_dummies(combined['SibSp'])

combined = pd.concat([combined, column_encoded], axis=1)

Train = combined[combined['train'] == 1]
Test = combined[combined['train'] == 0]

Train.drop(['train'], axis=1, inplace=True)
Test.drop(['train'], axis=1, inplace=True)

Train['train'] = 1
Test['train'] = 0

combined = pd.concat([Train, Test])

column_encoded = pd.get_dummies(combined['Parch'])

combined = pd.concat([combined, column_encoded], axis=1)

Train = combined[combined['train'] == 1]
Test = combined[combined['train'] == 0]

Train.drop(['train'], axis=1, inplace=True)
Test.drop(['train'], axis=1, inplace=True)

Train['train'] = 1
Test['train'] = 0

combined = pd.concat([Train, Test])

column_encoded = pd.get_dummies(combined['Embarked'])

combined = pd.concat([combined, column_encoded], axis=1)

Train = combined[combined['train'] == 1]
Test = combined[combined['train'] == 0]

Train.drop(['train'], axis=1, inplace=True)
Test.drop(['train'], axis=1, inplace=True)

Train['train'] = 1
Test['train'] = 0

combined = pd.concat([Train, Test])
column_encoded = pd.get_dummies(combined['Age_group'])
combined = pd.concat([combined, column_encoded], axis=1)
Train = combined[combined['train'] == 1]
Test = combined[combined['train'] == 0]
Train.drop(['train'], axis=1, inplace=True)
Test.drop(['train'], axis=1, inplace=True)

Train['train'] = 1
Test['train'] = 0
combined = pd.concat([Train, Test])
column_encoded = pd.get_dummies(combined['Cabin_New_2'])
combined = pd.concat([combined, column_encoded], axis=1)
Train = combined[combined['train'] == 1]
Test = combined[combined['train'] == 0]
Train.drop(['train'], axis=1, inplace=True)
Test.drop(['train'], axis=1, inplace=True)

Train['train'] = 1
Test['train'] = 0
combined = pd.concat([Train, Test])
column_encoded = pd.get_dummies(combined['Sex_New'])
combined = pd.concat([combined, column_encoded], axis=1)
Train = combined[combined['train'] == 1]
Test = combined[combined['train'] == 0]
Train.drop(['train'], axis=1, inplace=True)
Test.drop(['train'], axis=1, inplace=True)

Train['train'] = 1
Test['train'] = 0
combined = pd.concat([Train, Test])
column_encoded = pd.get_dummies(combined['Fare_group'])
combined = pd.concat([combined, column_encoded], axis=1)
Train = combined[combined['train'] == 1]
Test = combined[combined['train'] == 0]
Train.drop(['train'], axis=1, inplace=True)
Test.drop(['train'], axis=1, inplace=True)

Train.drop(['Fare_group', 'Pclass', 'SibSp', 'Parch', 'Embarked', 'Age_group', 'Cabin_New_2', 'Sex_New'], axis=1, inplace=True)

Test.drop(['Fare_group', 'Survived', 'Pclass', 'SibSp', 'Parch', 'Embarked', 'Age_group', 'Cabin_New_2', 'Sex_New'], axis=1, inplace=True)

Train.columns

Test.columns

Train = Train.rename(columns={'[60+)': 'A1', '[0, 20)': 'B1', '[20, 40)': 'C1', '[40, 60)': 'D1', '[31, 38.9)': 'E1', '[38.9+)': 'F1', '[7.9, 14.4)': 'G1', '[0, 7.9)': 'H1'})

Test = Test.rename(columns={'[60+)': 'A1', '[0, 20)': 'B1', '[20, 40)': 'C1', '[40, 60)': 'D1', '[31, 38.9)': 'E1', '[38.9+)': 'F1', '[7.9, 14.4)': 'G1', '[0, 7.9)': 'H1'})

Train.to_csv("Data/TrainEncoded.csv", index=False)
Test.to_csv("Data/TestEncoded.csv", index=False)
