==> 2.4- Data Preparation - Outliers, Normalization.ipynb <==
import pandas as pd
import numpy as np

Train =  pd.read_csv("Data/TrainEncoded.csv")
Test =  pd.read_csv("Data/TestEncoded.csv")

Train.head()

Train.shape

Test.head()

Test.shape

class Outliers:
    def __init__(self, df):
        self.df = df
    
    ## function to find outliers limits
    def find_limits_and_delete_outliers(self, variable):

        IQR = self.df[variable].quantile(0.75) - self.df[variable].quantile(0.25)

        lower_limit = self.df[variable].quantile(0.25) - (IQR * 1.5)
        upper_limit = self.df[variable].quantile(0.75) + (IQR * 1.5)

        ## to find and delete values upper and under the outliers limits
        np.where(self.df[variable] > upper_limit)
        np.where(self.df[variable] < lower_limit)
        
        outliers = np.where(self.df[variable] > upper_limit, True, 
                            np.where(self.df[variable] < lower_limit, True, False))
        
        outliers_df = self.df.loc[outliers, variable]
        outliers_df_ids = self.df.loc[outliers]
        new_df = self.df.drop(outliers_df_ids.index).reset_index(drop=True)
        
        outliers_percentage = len(outliers_df) / len(self.df)
        
        return outliers_percentage, new_df

outliers_instance = Outliers(Train)

print(Train.columns.tolist())

columns = ['Fare']

def removing_outliers(df, first_column, other_columns):
    outliers_instance = Outliers(df)
    outliers_percentage, new_df = outliers_instance.find_limits_and_delete_outliers(first_column)
    print(first_column, outliers_percentage, new_df.shape)
    
    for column in other_columns:
        outliers_instance = Outliers(new_df)
        outliers_percentage, new_df = outliers_instance.find_limits_and_delete_outliers(column)
        print(column, outliers_percentage, new_df.shape)
    return outliers_percentage, new_df

# outliers_percentage, new_df = removing_outliers(Train, columns[0], columns[1:])

new_df.shape

from sklearn.preprocessing import StandardScaler

class Normalization:
    def __init__(self, df):
        self.df = df
        
    def normalization(self, columns):
        scaler = StandardScaler()

        ## features to normalice
        normalice = self.df[columns]
        scaler.fit(normalice)
        df_scaled = scaler.transform(normalice)
        
        new_columns = [str(x)+'_scaled' for x in columns]
        
        df_scaled = pd.DataFrame(df_scaled, columns=new_columns)
        self.df = pd.concat([self.df,df_scaled], axis=1)
        
        self.df = self.df.drop(columns, axis=1)
        return self.df
    
    # Final dataframe
    def final_dataframe(self):
        return self.df

norm_instance_train = Normalization(new_df)

print(new_df.columns.tolist())

columns = ['Fare']

norm_instance_train.normalization(columns).head()

norm_instance_train.final_dataframe().head()

norm_instance_test = Normalization(Test)

norm_instance_test.normalization(columns).head()

norm_instance_train.final_dataframe().to_csv("Data/TrainOutliersNormalization.csv", index=False)
norm_instance_test.final_dataframe().to_csv("Data/TestOutliersNormalization.csv", index=False)
