==> 4- Baseline Models.ipynb <==
from sklearn.neighbors import KNeighborsClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.linear_model import RidgeClassifier
from sklearn.linear_model import SGDClassifier
from sklearn.linear_model import PassiveAggressiveClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.tree import ExtraTreeClassifier
from sklearn.svm import LinearSVC
from sklearn.svm import SVC
from sklearn.naive_bayes import GaussianNB
from sklearn.ensemble import AdaBoostClassifier
from sklearn.ensemble import BaggingClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import ExtraTreesClassifier
from sklearn.gaussian_process import GaussianProcessClassifier
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
from sklearn.discriminant_analysis import QuadraticDiscriminantAnalysis

import pandas as pd

Train = pd.read_csv('Data/TrainEncoded.csv')

Test = pd.read_csv('Data/TestEncoded.csv')

Train

Test

def get_datasets(train, test, target_var):
    y_train = Train[[target_var]]
    X_train = Train.drop(target_var, axis=1)
    X_test = Test
    print(len(X_train), len(X_train.columns), X_train.columns)
    return X_train, y_train, X_test

X_train, y_train, X_test = get_datasets(Train, Test, 'Survived')

def get_models():
    models = dict() #list() # append
    models.update({'KNeighborsClassifier': KNeighborsClassifier()})
    models.update({'LogisticRegression': LogisticRegression()})
    models.update({'RidgeClassifier': RidgeClassifier()})
    models.update({'SGDClassifier': SGDClassifier()})
    models.update({'PassiveAggressiveClassifier': PassiveAggressiveClassifier()})
    models.update({'DecisionTreeClassifier': DecisionTreeClassifier()})
    models.update({'ExtraTreeClassifier': ExtraTreeClassifier()})
    models.update({'LinearSVC': LinearSVC()})
    models.update({'SVC': SVC()})
    models.update({'GaussianNB': GaussianNB()})
    models.update({'AdaBoostClassifier': AdaBoostClassifier()})
    models.update({'BaggingClassifier': BaggingClassifier()})
    models.update({'RandomForestClassifier': RandomForestClassifier()})
    models.update({'ExtraTreesClassifier': ExtraTreesClassifier()})
    models.update({'GaussianProcessClassifier': GaussianProcessClassifier()})
    models.update({'GradientBoostingClassifier': GradientBoostingClassifier()})
    models.update({'LinearDiscriminantAnalysis': LinearDiscriminantAnalysis()})
    models.update({'QuadraticDiscriminantAnalysis': QuadraticDiscriminantAnalysis()})
    return models

get_models()

def results_to_csv(file_name, y_pred):
    model = pd.DataFrame({
        'PassengerId': list(range(1, len(y_pred)+1)),
        'Survived': y_pred
    })
    model.to_csv("results/{}.csv".format(file_name), index=False)
    return model

def modeling(train, test, target_var, num_iter):
    X_train, y_train, X_test = get_datasets(train, test, target_var)
    models = get_models()
    for key, value in models.items():
        clf = value.fit(X_train, y_train)
        y_pred = clf.predict(X_test).reshape(-1,)
        results_to_csv('{}-{}'.format(key, num_iter), y_pred)
    return y_pred

modeling = modeling(Train, Test, 'Survived', 1)

from xgboost import XGBClassifier

model = XGBClassifier(booster = 'gbtree', objective = 'binary:logistic', random_state = 2)

X_train = Train.drop('Survived', axis = 1)
y_train = Train[['Survived']]
X_test = Test

model.fit(X_train, y_train)

y_pred = model.predict(X_test)

y_pred

results_to_csv('XGBClassifier2', y_pred)

import catboost as cb
from catboost import CatBoostClassifier

model = CatBoostClassifier(verbose = 0, n_estimators = 100)

model

model.fit(X_train, y_train)

y_pred

results_to_csv('CatBoostClassifier', y_pred)
