==> 4.1 Resampling_renaming_models.ipynb <==
import pandas as pd
import numpy as np

df =  pd.read_csv("results/StackingClassifier.csv")

df

df['PassengerId2'] = df['PassengerId'].apply(lambda x: x+891)

df['Survived2'] = df['Survived'].apply(lambda x: int(x))

df

df = df.drop('PassengerId', axis=1)

df = df.drop('Survived', axis=1)

df

df = df.rename(columns={'PassengerId2': 'PassengerId', 'Survived2': 'Survived'})

df

df = df[["PassengerId", "Survived"]]

df.to_csv("results/StackingClassifier.csv", index=False)
