==> 7- Ensemble Learning.ipynb <==
from sklearn.neighbors import KNeighborsClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.svm import SVC
from sklearn.naive_bayes import GaussianNB
from sklearn.ensemble import AdaBoostClassifier ## Boosting
from sklearn.ensemble import RandomForestClassifier ## Ensemble/Bagging Model
from sklearn.gaussian_process import GaussianProcessClassifier
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis

import numpy as np
from matplotlib import pyplot as plt
import pandas as pd

X_oversample = pd.read_csv('Data/X_oversample.csv')
y_oversample = pd.read_csv('Data/y_oversample.csv')
Test = pd.read_csv('Data/TestEncoded.csv')

X_oversample

y_oversample

Test

def get_models():
    models = dict() #list() # append
    models.update({'KNeighborsClassifier': KNeighborsClassifier()})
    models.update({'LogisticRegression': LogisticRegression()})
    models.update({'DecisionTreeClassifier': DecisionTreeClassifier()})
    models.update({'SVC': SVC()})
    models.update({'GaussianNB': GaussianNB()})
    models.update({'AdaBoostClassifier': AdaBoostClassifier()})
    models.update({'RandomForestClassifier': RandomForestClassifier()})
    models.update({'GaussianProcessClassifier': GaussianProcessClassifier()})
    models.update({'LinearDiscriminantAnalysis': LinearDiscriminantAnalysis()})
    models.update({'StackingClassifier': get_stacking()})
    return models

from sklearn.ensemble import StackingClassifier

## crear el stacking de los modelos
def get_stacking():
    ## definir los modelos base
    level0 = list()
    level0.append(('knn', KNeighborsClassifier()))
    level0.append(('lr', LogisticRegression()))
    level0.append(('dt', DecisionTreeClassifier()))
    level0.append(('svm', SVC()))
    level0.append(('gnb', GaussianNB()))
    level0.append(('ab', AdaBoostClassifier()))
    level0.append(('rf', RandomForestClassifier()))
    level0.append(('gp', GaussianProcessClassifier()))
    level0.append(('ld', LinearDiscriminantAnalysis()))
    
    ## definir el metalearner
    level1 = LogisticRegression()
    
    ## definir el stacking-ensemble
    model = StackingClassifier(estimators=level0, final_estimator=level1, cv=5)
    return model

# from sklearn.model_selection import cross_val_score, KFold

# ## evluar el stacking usando corss validation
# def evaluate_model(model, X, y):
#     cv = KFold(n_splits=5)
#     scores = cross_val_score(model, X, y, scoring='f1', cv=cv, n_jobs=-1)
#     return scores

# models = get_models()

# results, names = list(), list()

# import numpy as np


# for name, model in models.items():
#     scores = evaluate_model(model, X_oversample, y_oversample)
#     results.append(scores)
#     names.append(name)
#     print(name, np.mean(scores), np.std(scores))
    
# plt.boxplot(results, labels=names, showmeans=True)
# plt.show()

from sklearn.datasets import make_classification

# def get_datasets():
#     X, y = make_classification(n_samples=1000, n_features=20, n_informative=15, 
#                               n_redundant=5, random_state=1)
#     return X, y

# X, y = get_datasets()

# X

def get_models():
    models = dict() #list() # append
    models.update({'SVC': SVC()})
    models.update({'RandomForestClassifier': RandomForestClassifier()})
    models.update({'GaussianProcessClassifier': GaussianProcessClassifier()})
    return models

from sklearn.model_selection import cross_val_score, KFold

## evluar el stacking usando corss validation
def evaluate_model(model, X_oversample, y_oversample):
    cv = KFold(n_splits=5)
    scores = cross_val_score(model, X, y, scoring='f1', cv=cv, n_jobs=-1)
    return scores

# models = get_models()
# results, names = list(), list()

# for name, model in models.items():
#     scores = evaluate_model(model, X, y)
#     results.append(scores)
#     names.append(name)
#     print(name, np.mean(scores), np.std(scores))
    
# plt.boxplot(results, labels=names, showmeans=True)
# plt.show()

##competicion

from sklearn.ensemble import StackingClassifier

## crear el stacking de los modelos
def get_stacking():
    ## definir los modelos base
    level0 = list()
    level0.append(('SVC', SVC()))
    level0.append(('RFC', RandomForestClassifier()))
    level0.append(('GPC', GaussianProcessClassifier()))

    ## definir el metalearner
    level1 = LogisticRegression() ## metalearner para un problema de clasificacion
    #level1 = LinearRegression() ## metalearner para un problema de regresion
    
    ## definir el stacking-ensemble
    model = StackingClassifier(estimators=level0, final_estimator=level1, cv=5)
    return model

def get_models():
    models = dict() #list() # append
    models.update({'SVC': SVC()})
    models.update({'RandomForestClassifier': RandomForestClassifier()})
    models.update({'GaussianProcessClassifier': GaussianProcessClassifier()})
    models.update({'StackingClassifier': get_stacking()})
    return models

models = get_models()
results, names = list(), list()

for name, model in models.items():
    scores = evaluate_model(model, X_oversample, y_oversample)
    results.append(scores)
    names.append(name)
    print(name, np.mean(scores), np.std(scores))
    
plt.boxplot(results, labels=names, showmeans=True)
plt.show()

from sklearn.ensemble import StackingClassifier

## crear el stacking de los modelos
def get_stacking():
    ## definir los modelos base
    level0 = list()
    level0.append(('SVC', SVC()))
    level0.append(('RFC', RandomForestClassifier()))
    level0.append(('GPC', GaussianProcessClassifier()))

    ## definir el metalearner
    level1 = LogisticRegression() ## metalearner para un problema de clasificacion
    #level1 = LinearRegression() ## metalearner para un problema de regresion
    
    ## definir el stacking-ensemble
    model = StackingClassifier(estimators=level0, final_estimator=level1, cv=5)
    return model

model = get_stacking()

model.fit(X_oversample, y_oversample)


y_pred = model.predict(Test)

y_pred

def results_to_csv(file_name, y_pred):
    model = pd.DataFrame({
        'PassengerId': list(range(1, len(y_pred)+1)),
        'Survived': y_pred
    })
    model.to_csv("results/{}.csv".format(file_name), index=False)
    return model

results_to_csv('StackingClassifier', y_pred)
